Revert "Explosions update: More items throwable, general refactoring (#34153)"
This reverts commit 17e24861d722cfa2d38b3ed62481e2e8199617c2.

+++ code/modules/mining/ore_box.dm
@@ -60,18 +60,22 @@
 	return
 
 /obj/structure/ore_box/proc/dump_everything()
-	. = list()
 	for(var/ore_id in stored_ores)
 		var/amount = stored_ores[ore_id]
 		if(amount > 0)
-			. += drop_stack(ore_id, get_turf(src), amount)
+			drop_stack(ore_id, get_turf(src), amount)
 
 	stored_ores.Cut()
 
 /obj/structure/ore_box/ex_act(severity)
-	if(prob(100-((severity-1)*50))) // 1= 100, 2 = 50, 3 = 0
-		. += dump_everything()
-		qdel(src)
+	switch(severity)
+		if(1.0)
+			dump_everything()
+			qdel(src)
+		if(2.0)
+			if (prob(50))
+				dump_everything()
+				qdel(src)
 
 /obj/structure/ore_box/proc/try_add_ore(var/obj/item/stack/ore/O)
 	if (!O.can_orebox)
